Move user data and client settings to context processor
So this data will be available on each request.
