+++ src/rf/rf/settings/base.py
@@ -197,6 +197,7 @@ TEMPLATE_CONTEXT_PROCESSORS = (
     'django.core.context_processors.tz',
     'django.contrib.messages.context_processors.messages',
     'django.core.context_processors.request',
+    'apps.core.context_processors.page_context',
 )
 
 # See: https://docs.djangoproject.com/en/dev/ref/settings/#template-loaders
